+++ apps/Standalone/src/templates/app/TemplatesStandard.tsx
@@ -53,7 +53,10 @@ export const TemplatesStandard = () => {
   }));
   const { appId, hostingPlan, workflowName: existingWorkflowName } = useSelector((state: RootState) => state.workflowLoader);
   const { data: workflowAppData } = useWorkflowApp(appId as string, hostingPlan);
-  const canonicalLocation = WorkflowUtility.convertToCanonicalFormat(workflowAppData?.location ?? '');
+  const canonicalLocation = useMemo(
+    () => WorkflowUtility.convertToCanonicalFormat(workflowAppData?.location ?? 'westus'),
+    [workflowAppData]
+  );
   const { data: tenantId } = useCurrentTenantId();
   const { data: objectId } = useCurrentObjectId();
   const { data: originalConnectionsData } = useConnectionsData(appId);
@@ -189,6 +192,7 @@ export const TemplatesStandard = () => {
   const services = useMemo(
     () =>
       getServices(
+        appId as string,
         connectionsData() ?? {},
         workflowAppData as WorkflowApp,
         addConnectionDataInternal,
@@ -197,11 +201,11 @@ export const TemplatesStandard = () => {
         canonicalLocation
       ),
     // eslint-disable-next-line react-hooks/exhaustive-deps
-    [connectionsData, settingsData, workflowAppData, tenantId, canonicalLocation]
+    [connectionsData, settingsData, workflowAppData, tenantId, canonicalLocation, appId]
   );
   const resourceDetails = new ArmParser(appId ?? '');
 
-  if (!workflowAppData) {
+  if (!appId) {
     return null;
   }
   return (
@@ -211,14 +215,38 @@ export const TemplatesStandard = () => {
           subscriptionId: resourceDetails.subscriptionId,
           resourceGroup: resourceDetails.resourceGroup,
           location: canonicalLocation,
-          workflowAppName: workflowAppData.name as string,
+          workflowAppName: resourceDetails.resourceName,
         }}
         connectionReferences={connectionReferences}
         services={services}
         isConsumption={false}
         isCreateView={true}
         existingWorkflowName={existingWorkflowName}
-        viewTemplate={isSingleTemplateView ? { id: templatesView } : undefined}
+        viewTemplate={
+          isSingleTemplateView
+            ? {
+                id: templatesView,
+                parametersOverride: {
+                  'odataTopDefault_#workflowname#': { value: 0, isEditable: false },
+                  'sharepoint-site-name_#workflowname#': { value: 'overriden-empty' },
+                  'TeamsChannelID_#workflowname#': { value: 'overriden-default', isEditable: false },
+                  'TeamsTeamID_#workflowname#': { value: 'overriden-default-editable' },
+                  'OpenAIEmbeddingModel_#workflowname#': { value: 'overriden-default-editable' },
+                  'SharepointSiteAddress_#workflowname#': { value: 'overriden-default-non-editable', isEditable: false },
+                },
+                basicsOverride: {
+                  [templatesView]: {
+                    name: { value: 'overriden-name', isEditable: false },
+                    kind: { value: 'stateful', isEditable: false },
+                  },
+                  ['ingest-index-ai-sharepoint-rag']: {
+                    name: { value: 'overriden-name', isEditable: false },
+                    kind: { value: 'stateful', isEditable: false },
+                  },
+                },
+              }
+            : undefined
+        }
       >
         <div
           style={{
@@ -285,6 +313,7 @@ const apiVersion = '2020-06-01';
 const httpClient = new HttpClient();
 
 const getServices = (
+  siteResourceId: string,
   connectionsData: ConnectionsData,
   workflowApp: WorkflowApp | undefined,
   addConnection: (data: ConnectionAndAppSetting) => Promise<void>,
@@ -292,11 +321,9 @@ const getServices = (
   objectId: string | undefined,
   location: string
 ): any => {
-  const siteResourceId = workflowApp?.id;
   const armUrl = 'https://management.azure.com';
   const baseUrl = `${armUrl}${siteResourceId}/hostruntime/runtime/webhooks/workflow/api/management`;
-  const appName = workflowApp?.name ?? '';
-  const { subscriptionId, resourceGroup } = new ArmParser(siteResourceId ?? '');
+  const { subscriptionId, resourceGroup, resourceName } = new ArmParser(siteResourceId ?? '');
 
   const defaultServiceParams = { baseUrl, httpClient, apiVersion };
 
@@ -311,7 +338,7 @@ const getServices = (
       tenantId,
       httpClient,
     },
-    workflowAppDetails: { appName, identity: workflowApp?.identity as any },
+    workflowAppDetails: { appName: resourceName, identity: workflowApp?.identity as any },
     readConnections: () => Promise.resolve(connectionsData),
     writeConnection: addConnection as any,
   });