+++ libs/designer/src/lib/core/actions/bjsworkflow/templates.ts
@@ -152,6 +152,28 @@ export const initializeTemplateServices = createAsyncThunk(
   }
 );
 
+export const loadManifestsFromPaths = async (resourcePaths: string[]) => {
+  try {
+    const manifestPromises = resourcePaths.map(async (resourcePath) => {
+      return import(`./../../templates/templateFiles/${resourcePath}/manifest.json`);
+    });
+    const manifestsArray = await Promise.all(manifestPromises);
+    return manifestsArray.reduce((result: Record<string, Template.Manifest>, manifestFile: any, index: number) => {
+      const manifest = manifestFile.default;
+      result[resourcePaths[index]] = manifest;
+      return result;
+    }, {});
+  } catch (error) {
+    LoggerService().log({
+      level: LogEntryLevel.Error,
+      area: 'Templates.loadGithubManifests',
+      message: `Error loading manifests: ${error}`,
+      error: error instanceof Error ? error : undefined,
+    });
+    return undefined;
+  }
+};
+
 export const loadTemplate = createAsyncThunk(
   'loadTemplate',
   async (
@@ -159,10 +181,12 @@ export const loadTemplate = createAsyncThunk(
     thunkAPI
   ) => {
     const currentState: RootState = thunkAPI.getState() as RootState;
-    const currentTemplateResourcePath = currentState.template.templateName;
+    const currentTemplateName = currentState.template.templateName;
+    const viewTemplateDetails = currentState.templateOptions.viewTemplateDetails;
+    const viewTemplateData = currentTemplateName === viewTemplateDetails?.id ? viewTemplateDetails : undefined;
 
-    if (currentTemplateResourcePath) {
-      return loadTemplateFromResourcePath(currentTemplateResourcePath, preLoadedManifest, isCustomTemplate);
+    if (currentTemplateName) {
+      return loadTemplateFromResourcePath(currentTemplateName, preLoadedManifest, isCustomTemplate, viewTemplateData);
     }
 
     return undefined;
@@ -203,7 +227,8 @@ export const validateWorkflowName = (workflowName: string | undefined, existingW
 const loadTemplateFromResourcePath = async (
   templateName: string,
   manifest: Template.Manifest | undefined,
-  isCustomTemplate: boolean
+  isCustomTemplate: boolean,
+  viewTemplateData?: Template.ViewTemplateDetails
 ): Promise<TemplatePayload> => {
   const templateManifest: Template.Manifest =
     manifest ?? (await import(`./../../templates/templateFiles/${templateName}/manifest.json`)).default;
@@ -227,7 +252,8 @@ const loadTemplateFromResourcePath = async (
         workflowPath,
         `${templateName}/${workflowPath}`,
         /* manifest */ undefined,
-        isCustomTemplate
+        isCustomTemplate,
+        viewTemplateData
       );
       if (workflowData) {
         workflowData.workflow.workflowName = workflows[workflowPath].name;
@@ -258,8 +284,9 @@ const loadTemplateFromResourcePath = async (
       }
     }
   } else {
-    const workflowId = 'default'; // This would change to the folder path which contains workflow manifest once we restructure.
-    const workflowData = await loadWorkflowTemplateFromManifest(workflowId, templateName, manifest, isCustomTemplate);
+    const workflowId = 'default';
+    const workflowData = await loadWorkflowTemplateFromManifest(workflowId, templateName, manifest, isCustomTemplate, viewTemplateData);
+
     if (workflowData) {
       data.workflows = {
         [workflowId]: workflowData.workflow,
@@ -276,7 +303,8 @@ const loadWorkflowTemplateFromManifest = async (
   workflowId: string,
   templatePath: string,
   manifest: Template.Manifest | undefined,
-  isCustomTemplate: boolean
+  isCustomTemplate: boolean,
+  viewTemplateData: Template.ViewTemplateDetails | undefined
 ): Promise<
   | {
       workflow: WorkflowTemplateData;
@@ -290,19 +318,26 @@ const loadWorkflowTemplateFromManifest = async (
     const parameterDefinitions = templateManifest.parameters?.reduce((result: Record<string, Template.ParameterDefinition>, parameter) => {
       result[parameter.name] = {
         ...parameter,
-        value: parameter.default,
+        value: viewTemplateData?.parametersOverride?.[parameter.name]?.value?.toString() ?? parameter.default,
         associatedWorkflows: [templateManifest.title],
       };
       return result;
     }, {});
+
+    const overridenKind = viewTemplateData?.basicsOverride?.[workflowId]?.kind?.value;
 
     return {
       workflow: {
         id: workflowId,
         workflowDefinition: (templateWorkflowDefinition as any)?.default ?? templateWorkflowDefinition,
         manifest: templateManifest,
-        workflowName: '',
-        kind: templateManifest.kinds?.length ? templateManifest.kinds[0] : 'stateful',
+        workflowName: viewTemplateData?.basicsOverride?.[workflowId]?.name?.value ?? '',
+        kind:
+          overridenKind && templateManifest.kinds?.includes(overridenKind)
+            ? overridenKind
+            : templateManifest.kinds?.length
+              ? templateManifest.kinds[0]
+              : 'stateful',
         images: templateManifest.images,
         connectionKeys: Object.keys(templateManifest.connections),
         errors: {