+++ libs/designer/src/lib/core/state/templates/templateselectors.ts
@@ -4,7 +4,7 @@ import type { WorkflowTemplateData } from '../../actions/bjsworkflow/templates';
 import type { ConnectionReference } from '../../../common/models/workflow';
 
 export const useAreServicesInitialized = () => {
-  return useSelector((state: RootState) => state.template.servicesInitialized ?? false);
+  return useSelector((state: RootState) => state.templateOptions.servicesInitialized ?? false);
 };
 
 export const useTemplateWorkflows = () => {
@@ -14,6 +14,15 @@ export const useTemplateWorkflows = () => {
 export const useWorkflowTemplate = (workflowId: string): WorkflowTemplateData => {
   return useSelector((state: RootState) => {
     return state.template.workflows[workflowId];
+  });
+};
+
+export const useWorkflowBasicsEditable = (workflowId: string) => {
+  return useSelector((state: RootState) => {
+    return {
+      isNameEditable: state.templateOptions.viewTemplateDetails?.basicsOverride?.[workflowId]?.name?.isEditable ?? true,
+      isKindEditable: state.templateOptions.viewTemplateDetails?.basicsOverride?.[workflowId]?.kind?.isEditable ?? true,
+    };
   });
 };